ajustado controler de GENEROS
