+++ database/seeders/GeneroSeeder.php
@@ -3,7 +3,6 @@
 namespace Database\Seeders;
 
 use App\Models\Genero;
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
 class GeneroSeeder extends Seeder
@@ -17,7 +16,7 @@ class GeneroSeeder extends Seeder
             ['nome'=>"VISUAL NOVEL"],
             ['nome'=>"ACTION RPG"],
             ['nome'=>"MOBA"],
-
+            ['nome'=>"SIMULADOR"]
         ]);
     }
 }
