+++ routes/api.php
@@ -1,13 +1,10 @@
 <?php
 
-use App\Http\Controllers\GeneroController;
+use App\Http\Controllers\Generos\GeneroApiController;
 use App\Http\Controllers\JogoController;
 use App\Http\Controllers\Login\LoginController;
 use Illuminate\Support\Facades\Route;
 
-// Route::get('/user', function (Request $request) {
-// return $request->user();
-// })->middleware('auth:sanctum');
 
 Route::group([
     "prefix" => "user",
@@ -22,11 +19,11 @@ Route::group([
     'prefix' => 'generos',
     'as' => 'generos.',
 ], function () {
-    Route::get('list', [GeneroController::class, 'getAll']);
-    Route::get('get/{id}', [GeneroController::class, 'getOne']);
-    Route::put('update/{id}', [GeneroController::class, 'updateOne']);
-    Route::post('create', [GeneroController::class,  'createOne']);
-    Route::delete('delete/{id}', [GeneroController::class, 'deleteOne']);
+    Route::get('list', [GeneroApiController::class, 'getAll']);
+    Route::get('get/{id}', [GeneroApiController::class, 'getOne']);
+    Route::put('update/{id}', [GeneroApiController::class, 'updateOne']);
+    Route::post('create', [GeneroApiController::class,  'createOne']);
+    Route::delete('delete/{id}', [GeneroApiController::class, 'deleteOne']);
 });
 
 Route::group([
